ECLGraph: Add support for Local Grid Refinement
This commit extends the initial implementation of the derivation of
an unstructured graph from an ECLIPSE result set to account for the
possibility of there being local grid refinement.  In particular, we
merge the existing classes

  ScatterMap, GlobalCellData, and CellCollection

into a new class, CartesianGridData, that supports generation of
Cartesian connections and extraction of data values on those
connections (e.g., phase flow rates) on one ERT grid--irrespective
of the nature of that grid (main grid or LGR).  The ECLGraph::Impl
class gains a vector of such CartesianGridData instances--one for
each active grid in the model.  We furthermore defer managing the
non-neighbouring connections to an internal class Impl::NNC that
queries the Cartesian grids when deriving the neighbourship
relations and extracting flux values.