ECLGraph: Expose mapping from Grid+(I,J,K) to active
This commit introduces a new public method

    activeCell(ijk, gridID = 0) const

in the ECLGraph class.  This method is useful for translating index
tuples that define a set of well completions to the linear, active
cell indices across all active grids in a result set.

The grid ID defaults to the main grid, but must be supplied if the
index tuples are relative to an LGR.