ECLGraph: Protect against missing NNC fluxes
This commit installs a guard in the handling of fluxes from
non-Cartesian connections (i.e., categories of NNC keywords).  In
particular we do not append NNC values to the global flux vector
unless all entries in the NNC vector have actually been extracted
from the relevant result set.

This fixes a problem in 'computeToFandTraces' when asking for gas
fluxes in an oil/water system.

Pointy hat: Bard.Skaflestad@sintef.no