Distinguish Restart Data from Init Data
This commit separates the original class ECLResultData into two
separate cases, ECLRestartData and ECLInitFileData.  The internal
structure of these result sets differs significantly in the presence
of local grids (i.e., LGRs) and we need to be aware of the
differences in order to interpret the INIT data correctly.

While here, switch to querying result set vectors in terms of grid
names rather than numerical grid IDs.  This makes the look-up
process more expensive, but also more resilient in case of possible
reordering of the local grids between result sets or data for local
grids being provided in multiple LGR sections.  Add a couple of
caches to reduce the impact of the more costly look-up process.