Add Support for Horizontal End-Point Scaling of Relperm
This commit introduces a facility for scaling the input values to
the one-dimensional interpolants according to ECL's two-point or
three-point scaling methods.  We also add latent support for scaling
the capillary pressure curve (two-point method, left end-point at
connate saturations), but this is currently not enabled in the
top-level, outer API.  Vertical scaling of relative permeabilities
is currently not supported.  Since the end-point scaling support
depends on availability of the tables' unscaled end-points, we add
query functions to class 'ECLPropTable1D' to retrieve these values.
This, in turn, begets a renaming to 'SatFuncInterpolant' to reflect
the more restricted applicability and scope of the resulting
interface.

We add a top-level management structure, class ECLSaturationFunc,
that collects one-dimensional interpolants for the O/G and O/W
subsystems and applies EPS if requested.  This class exposes a
single member function,

    kr = ECLSaturationFunc::relperm(G, rstrt, p)

that computes the relative permeabilities of active phase 'p' from
the result set 'rstrt' on all active cells of the model 'G'.  This
function is aware of ECL's saturation region mapping (SATNUM) and
calls the appropriate interpolants according to the available
mapping in a particular result set (.INIT file).  This function is
mainly intended for client code that needs to compute relative
permeability values--for instance to derive phase fluxes or to
visualize the relative permeability field as a 3D result vector.