computeLocalSolutions: Compute Local ToF and Tracers for All Wells
This commit extends the 'computeLocalSolutions' utility to compute
both time-of-flight and tracer concentrations for all wells in a
model that are active at a particular report step.  Time-of-flight
values are printed to files named

    tof-<wellname>.out

while tracer concentration values are printed to files named

    conc-<wellname>.out

While here, also add a bit of consistency checking to the ToF and
tracer solution result sets.  In particular, check that these
solutions have the same reachability in terms of cell IDs and print
a failure diagnostic if not.