Property Interpolants: Generalise Object Construction
This commit extracts the per-region interpolant construction loop
from the SatFuncInterpolant constructor out to a helper facility,

    Opm::MakeInterpolants<>::fromRawData()

This function is aware of the sub-structure imposed by the tabulated
live oil/wet gas property functions in the ECL Init result set and
knows how to position the column iterators appropriately for each
individual sub-table.  We call a factory function supplied by the
client for each such region, on the implicit assumption that this
constructor function knows how to advance the column iterators
across 'raw.numRows' entries.

All factory functions that are ultimately implemented in terms of

    Interp1D::PiecewisePolynomial::Linear<>

(or future facilities that have the same constructor semantics)
honour the requisite protocol.