PVT Interpolants: Add Estimated Derivative Extrapolation Policy
This commit adds a new extrapolation policy, Linearly, for the
piecewise linear interpolation class.  This policy estimates
function derivatives using table points in the "first" and "last"
intervals.  This is in contrast to the existing policy

    LinearlyWithDerivatives

that uses tabulated derivatives to extend a function definition
outside its range of input data.  Testing suggests that those
tabulated derivatives aren't particularly suited to extending the
range and that we get more plausible results by estimating the
derivatives ourselves.

Switch to using this new extrapolation policy for the WetGas and
LiveOil fluid models.