PVT Interpolants: Combine Reciprocal Quantities
This commit splits the evaluation of miscible PVT properties (class
PVTx) into two stages, an inner stage and an outer stage.  The inner
stage samples tabulated curves at both ends of an interval of the
primary lookup key and generates a convex combination of those
sampled values.  The outer stage performs any additional processing
needed to convert the convex combination into a scalar value of type
double.

In the case of the FVF, the inner stage computes the appropriate
convex combination of (1 / B_1) and (1 / B_2) while the outer stage
simply computes the reciprocal of that convex combination.  In the
case of the phase viscosity, the inner stage samples the pairs

    { 1/B, 1/(B*mu) }_1, { 1/B, 1/(B*mu) }_2

and computes the convex combination of those while the outer stage
computes the final viscosity value as

    (1 / B) / (1 / (B*mu))

from the convex combinations of those pairs.  Add a new helper
class, DenseVector<N>--implemented in terms of std::array<double,N>
to reuse the existing 2D interpolation engine when applied to a pair
of values.