Generalise Region Mapping Support
This commit introduces a new utility class, ECLRegionMapping, that
simplifies working with integer region mappings as implied by ECL
result set vectors such as SATNUM or PVTNUM.  The existing class
AssembledConnections from the FD engine library performs the heavy
lifting, but this class adds a layer of indirection that enables
using a subset--or a multiply sampled subset--of the mapping
vectors.

Add a unit test to demonstrate usage.