ECL Flux Calculator: Incorporate Viscosity and Gravity
This commit makes the ECLFluxCalc facility aware of the effects of
variable phase densities, phase viscosities and gravity.  In
particular we now compute different dynamic data--revised to be

    phase pressure
    phase mobility
    phase density

for each active fluid phase (at reservoir conditions).  The density
and mobility calculations for gas and oil are aware of component
mixing behaviour (wet gas w/vaporised oil, live oil w/dissolved gas)
and will include those effects if the requisite data ("RV" and "RS",
respectively) is availble in the result set.

The water calculations are simpler, and use only the traditional
second order Taylor approximation to the reciprocal FVF and (B*mu)
functions--assuming constant compressibility and constant
viscosibility.

At present we do not include effects of capillary pressure, so all
fluid properties are sampled in the oil pressure value of the cell.