Add Utility for Centralising ECL Result Set File Handling
This commit introduces a new class, ResultSet, that when initialized
from either a case-name prefix or a path to one of the typical
output files (.EGRID, .UNRST &c) will collect the paths that
represent all output files of that case.

We support both unified and separate restart files, and provide a
way of extracting all recorded/saved report steps from the RSSPEC
file.