Saturation Function: Add Protocol for Extracting Curves
This commit adds a new member function

    ECLSaturationFunc::getSatFuncCurve()

that enables retrieving the tabulated or effective curves of a
sequence of one or more saturation functions--represented as
FlowDiagnostics::Graph objects.  At present we support the relative
permeability functions of the basic two-dimensional sub-systems
(i.e., O/G and O/W).

This facility is aware of end-point scaling and will allow retrieval
of the raw curves or the (horizontally) scaled, effective curves
based on a run-time parameter.  Scaling mode for rel-perm is not
configurable, but uses the INIT file's setting of item 20 of the
LOGIHEAD vector (activates alternative/three-point scaling).