PVT Curves: Support Extracting Saturated State Envelope
This commit introduces a new PVT curve category, SaturatedState,
that represents the curve of saturated states (RvSat(Pg) and
RsSat(Po)).  These are useful in their own right and the former is
more or less required in order to produce a graphical representation
of wet gas properties in a diagram versus gas pressure.