PVT Curves: Clarify Distinction Between "Main" and "Saturated"
This commit introduces a new member function, mainPvtCurve(), in the
WetGas and LiveOil implementation classes that helps to distinguish
the "main" PVT curves (i.e., formation volume factor and viscosity)
from the saturated state curve (RvSat(Pg) or RsSat(Po)).  This
distinction reenables the curve copy elision property that was lost
in commit eaac39e and also serves as an opportunity to fix a
shadowing warning ('curve' identifier).

Thanks to [at]atgeirr for running this through CLang and helping to
identify the issues.