Saturation Functions: Return KroX as Function of Sx
This commit introduces a slight transformation of the relative
permeability functions for oil (i.e., Krog and Krow) to ensure that
these are given as functions of the corresponding "other" phase
saturation (i.e., Krog = Krog(Sg) and Krow = Krow(Sw)) rather than
as functions of oil saturation.

These functions are *stored* in the INIT result set as functions of
So, but it is customary to present them as functions of Sg or Sw in
the context a graphical outputs.