ECL PVT: Add Facility for Calculating Props per Cell
This commit adds a new member function,

    ECLPvtCurveCollection::getDynamicProperty()

that enables calculating prinicpal dynamic properties like the phase
formation volume factor or the phase viscosity in a single cell at a
collection of cell states (phase pressure and mixing ratios).  One
example of such related cell states would be the entire time series
of an ECL result set.  We reuse the "RawCurve" description to
identify which property to calculate and simply forward to the
corresponding member functions of the fundamental Oil and Gas
property interpolants.