ECL PVT: Add Example to Demonstrate Dynamic Property Calculation
This commit adds a new example utility that computes a sequence of
dynamic property values (subset of "Bg", "mu_g", "Bo", "mu_o") in a
single active cell based on a result set's entire time series of
phase pressures and phase mixing ratios for that cell.  The result
is printed as a sequence of Octave/MATLAB-compatible variable
assignments for easy loading into those environments.

Example invocation:

    ${exe} init=path/to/Case.INIT \
        cell=31415 mu_g=true Bg=true > props_31415.m