PVT Curves: Clarify API Unit Handling Responsibilities
This commit separates the unit handling concerns from the property
calculation concerns of class ECLPvtCurveCollection.  In particular
we now allow the caller to identify whether or not to use the native
(i.e., serialised) units or strict SI units when computing a dynamic
property (formation volume factor or phase viscosity).  Moreover,
the client code can specify a particular unit system convention for
results and this convention need not coincide with either of the
serialised or internal unit conventions.

In particular we introduce a new member function

    void setOutputUnits(unique_ptr<const UnitSystem> usys)

that allows the client to specify a collection of units for results
that are not in the strict SI system of units.  This member function
affects the outputs from

    vector<Graph> getPvtCurve()

and the new member function

    vector<double> getDynamicPropertyNative()

The inputs to the latter are assumed to be in the native/serialised
units of measure (e.g., in FIELD units if the result set is stored
as FIELD).  In other words getDynamicPropertyNative() assumes that
it is given quantities that have been extracted directly from an ECL
result set (typically a restart/report set) without any unit
conversion.  Its results will be returned in the unit convention
defined by setOutputUnits(), or strict SI if setOutputUnits() has
not been called.

Finally, the new member function

    vector<double> getDynamicPropertySI()

assumes that its inputs are in strict SI units and it will produce
results that are in strict SI units.

Update callers accordingly, and simplify the interpolants since they
no longer need to be concerned with unit conversion of results.