Endpoint Scaling: Guard Against Nonsensical Scaled Saturations
Real-life testing has demonstrated that values of scaled endpoints
need not always have reasonable values (i.e., in the range [0, 1]).
One particular case has scaled saturations of -1.0E+20 which lead to
very peculiar values--for instance in a graphical representation.

This commit introduces a user-configurable run-time policy,
implemented as a constructor argument to the EPS classes, called

    EPSEvalInterface::InvalidEndpointBehaviour

with possible values "UseUnscaled" and "IgnorePoint".  The former
policy means that invalid scaled saturations get translated to the
unscaled saturation while the latter means that cells with invalid
scaled saturations are ignored.  The first of these is the default
setting and is typically appropriate for simulation applications
while the latter may be useful for graphical output of saturation
function curves.