EPS: Support Three-Point Option in Two-Phase Systems
The original extraction of the unscaled table end-points for the
three-point relative permeability scaling option failed to take into
account that there might not always be three active phases.  This
commit adds the requisite awareness--at least for G/O and O/W
systems.  In particular, we assign unscaled minimum gas or water
saturation to zero depending on phase presence.

Live testing is needed to determine if this solves the problem
mentioned by [at]vkip.