Build Prerequisites: Retire -Core, Introduce -Parser
We used to use opm-core for its parameter system (ParameterGroup
and related features).  This system now resides in opm-common so we
get it for free.  Retire opm-core as a prerequisite, but add
opm-parser for its unit conversion system.  The parser module was
previously added transitively as a prerequisite for opm-core.