Sat-func Scaling: Generalize Option Specification
This commit introduces a new system for specifying saturation
function scaling options.  In particular, we introduce a new option
structure, ECLSaturationFunc::SatFuncScaling, that holds a bit
patterns govering which options to enable (only Horizontal supported
at this time) and how to treat "invalid" scaled saturations.

We now, furthermore, always enable saturation function scaling upon
constructing an ECLSaturationFunc object and leave the decision for
whether or not to actually scale the curves until
runtime--especially for the ECLSaturationFunc::getSatFuncCurve()
member function.

Update in-tree callers accordingly, but keep the existing overload
to function getSatFuncCurve() (with a default 'bool' parameter) for
backwards compatiblity with out-of-tree callers.  Re-implement the
existing interface in terms of the new function overload.