ECLGraph: Add Method for Extracting Vector from Subgrid
This commit adds a new way of extracting a grid vector only from a
subgrid.  This is useful if the calling context knows that the
vector may be defaulted (not present) and has additional information
for how to reconstruct the grid vector in this case.

This is needed to handle some corner cases of the KR* vectors.