EPS: Use Correct Maximum So for Converting to "S_other"
This commit corrects a mistake in translating oil saturation to
"other" saturation values in two-phase (G/O or O/W) systems.  We
must, of course, use the scaled maximum oil saturation in the
relevant 2p-system when doing the subtraction rather than simply the
maximum oil saturation in the SOFN table.