Flux and massflux: Return Empty Vector for Inactive Phases
This commit introduces a new (private) predicate function

    ECLFluxCalc::phaseIsActive(const ECLPhaseIndex phase) const

that identifies whether or not 'phase' is active in the current
simulation model/result set.  This enables returning an empty result
vector from flux() and/or massflux() in the case of the caller
requesting fluxes for inactive phases.  That, in turn, mirrors the
behaviour of ECLGraph::flux() on inactive phases and promotes
uniform phase treatment in calling code irrespective of using stored
or calculated flux values.