Vertical Scaling/Sr: Remove A Few Asserts
These were meant for consistency checking, but would occasionally
fail due to round-off error when doing arithmetic on values
converted from 'float' to 'double'.  The actual scaling procedure
does not rely on those relations as long as we ensure that 'fmax' is
the maximum function value, so add a measure of robustness here.

Note that this is at best a work-around for a deeper problem and
that a more refined solution would probably be needed here.