Vertical Scaling [PCOW]: Use Value at Minimum Saturation
The oil-water capillary pressure is a non-increasing function of
water saturation so we need to ensure that we use the capillary
pressure value at the minimum tabulated water saturation.  Add
special purpose code to enforce this rule.  This adds a certain
amount of computational overhead because we now compute function
values at two saturation points instead of just one for the case of
pure vertical scaling.  Most of the time those extra function values
will just be subsequently discarded.

Note that this is a bit of hack, because it relies on the fact that
the current implementation assigns

  TableEndPoints::disp = TableEndPoints::low

in the case of two-point horizontal scaling.  We may wish to make
that rule more explicit.